Move macros and plugin_main to other files
